+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the "vscode-run-function" extension will be documented in
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-11-12
+
+### Added
+
+-   Added the `run-function.buttonDisplay` configuration to define how to display the button to run the function
+
+### Deprecated
+
+-   Deprecated the `run-function.enableCodeLens` configuration in favor of the new `run-function.buttonDisplay` configuration. This will be removed on version 1
+
 ## [0.1.1] - 2021-11-08
 
 ### Fixed